+++ app.py
@@ -6,15 +6,14 @@ from dash import dcc, html
 from dash.dependencies import Input, Output
 import dash_bootstrap_components as dbc
 import pandas as pd
-
+import torch 
+from scripts.utils import LSTMRegressor, device, learning_rate, n_hidden_units
 ################ DATA PREPROCESSING ################
 
-model=joblib.load('d:/Documents-folders/GitHub/BoozAllen/model2140.joblib')
 
-
+# Read data
 df_test = pd.read_csv("d:/Documents-folders/GitHub/BoozAllen/CMAPSSData/test_FD001.txt", header=None, sep = ' ')
 df_train = pd.read_csv("d:/Documents-folders/GitHub/BoozAllen/CMAPSSData/train_FD001.txt", header=None, sep = ' ')
-
 
 
 ## Refactor data wrangling commands
@@ -44,7 +43,16 @@ df_test=smoothing(df_test)
 df_train=drop_org(df_train)
 df_test=drop_org(df_test)
 
-y = model.predict(df_test.iloc[0:1,2:])
+# Instantiate the model
+n_features = len([c for c in df_train.columns if 's' in c])
+loaded_model = LSTMRegressor(n_features, n_hidden_units)
+
+# Load the saved state_dict
+model_path = "model2140_1.pt"
+loaded_model.load_state_dict(torch.load(model_path))
+
+
+y = loaded_model.predict(df_test.iloc[0:1,2:])
 ############# APP LAYOUT #############
 
 external_stylesheets = ['https://codepen.io/chriddyp/pen/bWLwgP.css']
